Fix detection of java for flatpak
The logic for fdroidserver to find java is defined here:
https://gitlab.com/fdroid/fdroidserver/blob/0.8/fdroidserver/common.py#L154

For me it was not possible to match the regex as I wasn't able to move
the directory of java in flatpak. All I can do are symlinks, but the
fdroidserver code checks whether a given JAVA_HOME is a symlink and then
drops that path.

This is the best solution I came up with, which works now. In future, I
want to update fdroidserver directly to allow flatpak installations in
its regex, but as we don't want to bump fdroidserver before 1.0.0, I
decided to go this way.