also support using fdroidserver 2.1
sha256sum() is too trivial to deal with getting it from the private
fdroidserver API, and it moved from update to common.  Plus, it could be
useful to have it here so that the it can be hooked into any other
process that might be also reading the contents of the APK file.